Add implementation of Q-learning using function approximation.
A single Fully-connected layer is being used as the function approximator. The task this time is the MountainCar. Up to this point, the agent does not seem to learn. I suspect this could be due to the approximator being used, that might not be good enough. Moreover, I have just run the learning for 1000 episodes.